:sparkles:<feat, fix>(Makefile) : add Makefile, fix error
Commit type:

Description :

Motivation:

+++ ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: rnaka <rnaka@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/17 10:40:08 by rnaka             #+#    #+#             */
-/*   Updated: 2023/11/18 22:42:29 by rnaka            ###   ########.fr       */
+/*   Updated: 2023/11/19 20:46:29 by rnaka            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,13 +26,13 @@ int	main(int argc, char *argv[])
 		std::cout << "* LOUD AND UNBEARABLE FEEDBACK NOISE *" << std::endl;
 		return (0);
 	}
-	for (size_t i = 1; i < argc; i++)
+	for (size_t i = 1; static_cast<int>(i) < argc; i++)
 	{
 		for (size_t j = 0; j < std::strlen(argv[i]); j++)
 		{
 			std::cout << static_cast<char>(std::toupper(argv[i][j]));
 		}
-		if (i != argc - 1)
+		if (static_cast<int>(i) != argc - 1)
 			std::cout << ' ';
 	}
 	std::cout << std::endl;
